fix(query-engine): restore move-based single-accumulator path, dedupe test fixtures
Two follow-ups from review of the #596 fix:

- merge_accumulators_batch always cloned the first accumulator, even
  for the common single-bucket case, regressing the move-based
  shortcut SimpleEngine::merge_accumulators had before this refactor.
  Restore it at the call site (NaiveMerger can't do the same since it
  merges from a borrowed Vec it keeps for the next slide()).
- The new regression tests duplicated PoisonableAccumulator,
  cms_from_matrix, and oracle_sequential_fold verbatim between
  window_merger.rs and simple_engine/mod.rs; moved them into
  tests::test_utilities::merge_fixtures, shared by both.
